feat: move sendResearcherCloseoutCompletedMessage
